// src/app/pages/hotel/view-details-hotel/shared/view-details-hotel-enum.ts
export enum DetailsHotelEnum {
    title= 'Property Preview',
    // tslint:disable-next-line:max-line-length
    subtitle = 'This is an overview of all the information you\'ve provided about your property. You can edit or update these details and preview what your property might look like on ZOCO.',
    id = 'idItem',

    // fields
    types = 'Room types at',
    rooms = 'Number of rooms: ',
    policies = 'Policies',

    // Columns
    max_occupancy = 'max_occupancy',
    price = 'price',
    room_type = 'room_type',
    room_name = 'room_name',
    number_of_this_room = 'number_of_this_room',
    living_quantity = 'living_quantity',
    bath_quantity = 'bath_quantity',
    size = 'size',
    measure = 'measure',
    unitOfMeasurement = 'unitOfMeasurement',
    // rows
    type = 'Type',
    name = 'Name',
    number = 'Quantity',
    living = 'Living quantity',
    bath = 'Bath quantity',
    sizeTitle = 'Size',
    occupancy = 'Max occupancy',
    priceTitle = 'Price',
    measureTitle = 'Measure',

    // policies
    check_in = 'Check in',
    check_out = 'Check out',

    // carrousel
    slyde = 'ngb-slide-',

    getDetailHotelErrorMessage = 'There was an error trying to get Hotel Details',



    idDefaultValue= 1,
    establishmentNameDefaultValue= 'Loading Establishment Name..',
    countryDefaultValue= 'Loading Country..',
    homeAddressDefaultValue= 'Loading Home Address..',
    quantityRoomsDefaultValue= 1,
    starRatinDefaultValue= 0,
    propertyDescriptionDefaultValue= 'Loading Property Description..',
    hotelEstablishmentTypeValue = 11,
    maxGuestIconToShow = 3,
    noStarValue = 0,
    start = 0,
}